Re #216 - move to the addie.processing.mantid.launch_reduction.py location
