+++ addie/processing/mantid/event_handler.py
@@ -1,7 +1,4 @@
-import os
-import traceback
 from qtpy.QtWidgets import QFileDialog
-from mantidqt.utils.asynchronous import AsyncTask
 
 from addie.processing.mantid.master_table.table_tree_handler import TableTreeHandler, TableConfig, H3TableHandler
 from addie.processing.mantid.master_table.selection_handler import TableHandler as MtdTableHandler
@@ -14,7 +11,7 @@ from addie.processing.mantid.master_table.import_from_database.load_into_master_
 from addie.processing.mantid.make_calibration_handler.make_calibration import MakeCalibrationLauncher
 from addie.processing.mantid.master_table.reduction_configuration_handler import ReductionConfigurationHandler
 from addie.processing.mantid.master_table.master_table_loader import AsciiLoader
-from addie.processing.mantid.master_table.master_table_exporter import TableFileExporter as MantidTableExporter
+
 from addie.processing.mantid.master_table.tree_definition import INDEX_OF_COLUMNS_WITH_MASS_DENSITY
 
 # ONCat integration
@@ -25,96 +22,6 @@ except ImportError:
     print('pyoncat module not found. Functionality disabled')
     ONCAT_ENABLED = False
 
-# Mantid Total Scattering integration
-# (https://github.com/marshallmcdonnell/mantid_total_scattering)
-try:
-    import total_scattering
-    print("Mantid Total Scattering Version: ", total_scattering.__version__)
-    from total_scattering.reduction import TotalScatteringReduction
-    MANTID_TS_ENABLED = True
-except ImportError:
-    print('total_scattering module not found. Functionality disabled')
-    MANTID_TS_ENABLED = False
-
-
-class JobPool(object):
-    task_output = None,
-    running = None
-    task_exc_type, task_exc, task_exc_stack = None, None, None
-
-    def __init__(self, configurations):
-        self.jobs = []
-        for config in configurations:
-            print("CONFIG:", config)
-            self.jobs.append(AsyncTask(TotalScatteringReduction, args=(config,),
-                                       success_cb=self.on_success, error_cb=self.on_error,
-                                       finished_cb=self.on_finished))
-
-    def _start_next(self):
-        if self.jobs:
-            self.running = self.jobs.pop(0)
-            self.running.start()
-        else:
-            self.running = None
-
-    def start(self):
-        if not self.jobs:
-            raise RuntimeError('Cannot start empty job list')
-        self._start_next()
-
-    def on_success(self, task_result):
-        # TODO should emit a signal
-        self.task_output = task_result.output
-        print('SUCCESS!!! {}'.format(self.task_output))
-
-    def on_error(self, task_result):
-        # TODO should emit a signal
-        print('ERROR!!!')
-        self.task_exc_type = task_result.exc_type
-        self.task_exc = task_result.exc_value
-        self.task_exc_stack = traceback.extract_tb(task_result.stack)
-        traceback.print_tb(task_result.stack)
-
-    def on_finished(self):
-        '''Both success and failure call this method afterwards'''
-        # TODO should emit a signal
-        self._start_next()  # kick off the next one in the pool
-
-
-def run_mantid(parent):
-    num_rows = parent.processing_ui.h3_table.rowCount()
-    if num_rows <= 0:
-        raise RuntimeError('Cannot export empty table')
-
-    exporter = MantidTableExporter(parent=parent)
-
-    # write out the full table to disk
-    # TODO make a class level name so it can be reused
-    full_reduction_filename = os.path.join(
-        os.path.expanduser('~'), '.mantid', 'addie.json')
-    print('writing out full table to "{}"'.format(full_reduction_filename))
-    exporter.export(full_reduction_filename)
-
-    # append the individual rows to input list (reduction_inputs)
-    reduction_inputs = []
-    for row in range(num_rows):
-        if not exporter.isActive(row):
-            print('skipping row {} - inactive'.format(row + 1))  # REMOVE?
-            continue
-        print('Will be running row {} for reduction'.format(
-            row + 1))  # TODO should be debug logging
-        json_input = exporter.retrieve_row_info(row)
-        reduction_input = exporter.convert_from_row_to_reduction(json_input)
-        reduction_inputs.append(reduction_input)
-    if len(reduction_inputs) == 0:
-        raise RuntimeError('None of the rows were activated')
-
-    # locate total scattering script
-    if MANTID_TS_ENABLED:
-        pool = JobPool(reduction_inputs)
-        pool.start()
-    else:
-        print('total_scattering module not found. Functionality disabled') # TODO should be on the status bar
 
 def personalization_table_clicked(main_window):
     _ = TableTreeHandler(parent=main_window)
@@ -152,17 +59,20 @@ def load_this_config(main_window, key='', resize=False):
     for _col in h1_dict:
         _visible = h1_dict[_col]['visible']
         _width = h1_dict[_col]['width']
-        o_table.set_size_and_visibility_column(h1=_col, width=_width, visibility=_visible, resize=resize)
+        o_table.set_size_and_visibility_column(
+            h1=_col, width=_width, visibility=_visible, resize=resize)
 
     for _col in h2_dict:
         _visible = h2_dict[_col]['visible']
         _width = h2_dict[_col]['width']
-        o_table.set_size_and_visibility_column(h2=_col, width=_width, visibility=_visible, resize=resize)
+        o_table.set_size_and_visibility_column(
+            h2=_col, width=_width, visibility=_visible, resize=resize)
 
     for _col in h3_dict:
         _visible = h3_dict[_col]['visible']
         _width = h3_dict[_col]['width']
-        o_table.set_size_and_visibility_column(h3=_col, width=_width, visibility=_visible, resize=resize)
+        o_table.set_size_and_visibility_column(
+            h3=_col, width=_width, visibility=_visible, resize=resize)
 
     o_table.update_tree_dict_and_tree(config_to_load)
 
@@ -213,8 +123,10 @@ def resizing_h1(main_window, index_column, old_size, new_size):
         o_table.set_size_column(h1=index_column, width=old_size)
     else:
         last_h2_visible_size = o_table.get_size_column(h2=last_h2_visible)
-        o_table.set_size_column(h2=last_h2_visible, width=last_h2_visible_size + size_diff)
-        o_table.set_size_column(h3=last_h3_visible, width=last_h3_visible_size + size_diff)
+        o_table.set_size_column(
+            h2=last_h2_visible, width=last_h2_visible_size + size_diff)
+        o_table.set_size_column(
+            h3=last_h3_visible, width=last_h3_visible_size + size_diff)
 
     o_table.disconnect_table_ui(unblock_all=True)
 
@@ -237,7 +149,8 @@ def resizing_h2(main_window, index_column, old_size, new_size):
         # add this size_diff to parent and last h3
         parent_size = o_table.get_size_column(h1=h1_parent)
         o_table.set_size_column(h1=h1_parent, width=parent_size + size_diff)
-        o_table.set_size_column(h3=last_h3_visible, width=last_h3_visible_size + size_diff)
+        o_table.set_size_column(
+            h3=last_h3_visible, width=last_h3_visible_size + size_diff)
 
     o_table.disconnect_table_ui(unblock_all=True)
 
@@ -261,7 +174,8 @@ def resizing_h3(main_window, index_column, old_size, new_size):
 
 def init_tree(main_window):
     main_window.addItems(main_window.ui.treeWidget.invisibleRootItem())
-    main_window.ui.treeWidget.itemChanged.connect(main_window.tree_item_changed)
+    main_window.ui.treeWidget.itemChanged.connect(
+        main_window.tree_item_changed)
 
 
 def tree_item_changed(main_window, item):
@@ -272,7 +186,8 @@ def tree_item_changed(main_window, item):
                                   block_h1=True,
                                   block_h2=True)
 
-    h_columns_affected = o_table.get_h_columns_from_item_name(item_name=o_table.get_item_name(item))
+    h_columns_affected = o_table.get_h_columns_from_item_name(
+        item_name=o_table.get_item_name(item))
 
     o_table.change_state_tree(list_ui=h_columns_affected['list_tree_ui'],
                               list_parent_ui=h_columns_affected['list_parent_ui'],
@@ -289,7 +204,7 @@ def master_table_select_state_changed(main_window, state, key):
     o_table.activated_row_changed(key=key, state=state)
 
 
-## sample columns
+# sample columns
 def sample_material_button_pressed(main_window, key):
     MaterialHandler(parent=main_window, key=key, data_type='sample')
 
@@ -306,7 +221,8 @@ def sample_mass_density_button_pressed(main_window, key):
 def sample_mass_density_line_edit_entered(main_window, key):
     o_table = TableRowHandler(main_window=main_window)
     o_table.transfer_widget_states(from_key=key, data_type='sample')
-    main_window.check_master_table_column_highlighting(column=INDEX_OF_COLUMNS_WITH_MASS_DENSITY[0])
+    main_window.check_master_table_column_highlighting(
+        column=INDEX_OF_COLUMNS_WITH_MASS_DENSITY[0])
 
 
 def sample_shape_changed(main_window, index, key):
@@ -321,12 +237,14 @@ def sample_abs_correction_changed(main_window, text, key):
 
 def sample_multi_scattering_correction_changed(main_window, text, key):
     o_table = TableRowHandler(main_window=main_window)
-    o_table.multi_scattering_correction(value=text, key=key, data_type='sample')
+    o_table.multi_scattering_correction(
+        value=text, key=key, data_type='sample')
 
 
 def sample_inelastic_correction_changed(main_window, text, key):
     o_table = TableRowHandler(main_window=main_window)
-    o_table.inelastic_correction_changed(value=text, key=key, data_type='sample')
+    o_table.inelastic_correction_changed(
+        value=text, key=key, data_type='sample')
 
 
 def sample_placzek_button_pressed(main_window, key):
@@ -335,7 +253,8 @@ def sample_placzek_button_pressed(main_window, key):
 
 
 def sample_dimensions_setter_button_pressed(main_window, key):
-    o_dimensions_ui = DimensionsSetter(parent=main_window, key=key, data_type='sample')
+    o_dimensions_ui = DimensionsSetter(
+        parent=main_window, key=key, data_type='sample')
     o_dimensions_ui.show()
 
 
@@ -356,7 +275,8 @@ def normalization_mass_density_button_pressed(main_window, key):
 def normalization_mass_density_line_edit_entered(main_window, key):
     o_table = TableRowHandler(main_window=main_window)
     o_table.transfer_widget_states(from_key=key, data_type='normalization')
-    main_window.check_master_table_column_highlighting(column=INDEX_OF_COLUMNS_WITH_MASS_DENSITY[1])
+    main_window.check_master_table_column_highlighting(
+        column=INDEX_OF_COLUMNS_WITH_MASS_DENSITY[1])
 
 
 def normalization_shape_changed(main_window, text, key):
@@ -366,17 +286,20 @@ def normalization_shape_changed(main_window, text, key):
 
 def normalization_abs_correction_changed(main_window, text, key):
     o_table = TableRowHandler(main_window=main_window)
-    o_table.abs_correction_changed(value=text, key=key, data_type='normalization')
+    o_table.abs_correction_changed(
+        value=text, key=key, data_type='normalization')
 
 
 def normalization_multi_scattering_correction_changed(main_window, text, key):
     o_table = TableRowHandler(main_window=main_window)
-    o_table.multi_scattering_correction(value=text, key=key, data_type='normalization')
+    o_table.multi_scattering_correction(
+        value=text, key=key, data_type='normalization')
 
 
 def normalization_inelastic_correction_changed(main_window, text, key):
     o_table = TableRowHandler(main_window=main_window)
-    o_table.inelastic_correction_changed(value=text, key=key, data_type='normalization')
+    o_table.inelastic_correction_changed(
+        value=text, key=key, data_type='normalization')
 
 
 def normalization_placzek_button_pressed(main_window, key):
@@ -385,7 +308,8 @@ def normalization_placzek_button_pressed(main_window, key):
 
 
 def normalization_dimensions_setter_button_pressed(main_window, key):
-    o_dimensions_ui = DimensionsSetter(parent=main_window, key=key, data_type='normalization')
+    o_dimensions_ui = DimensionsSetter(
+        parent=main_window, key=key, data_type='normalization')
     o_dimensions_ui.show()
 
 
@@ -406,10 +330,10 @@ def make_calibration_clicked(main_window):
 
 def browse_calibration_clicked(main_window):
     _calibration_folder = main_window.calibration_folder
-    [_calibration_file, _] = QFileDialog.getOpenFileName(parent = main_window,
-                                                         caption = "Select Calibration File",
-                                                         directory = _calibration_folder,
-                                                         filter = main_window.calibration_extension)
+    [_calibration_file, _] = QFileDialog.getOpenFileName(parent=main_window,
+                                                         caption="Select Calibration File",
+                                                         directory=_calibration_folder,
+                                                         filter=main_window.calibration_extension)
     if _calibration_file:
         main_window.processing_ui.calibration_file.setText(_calibration_file)
